+++ AllMyLights.Test/Connectors/WallpaperSinkTest.cs
@@ -1,8 +1,9 @@
+using System.Collections.Generic;
 using System.IO;
 using AllMyLights.Connectors.Sinks.Wallpaper;
-using AllMyLights.Models.OpenRGB;
 using AllMyLights.Platforms;
 using Moq;
+using Newtonsoft.Json.Linq;
 using NUnit.Framework;
 
 namespace AllMyLights.Test
@@ -19,6 +20,26 @@ namespace AllMyLights.Test
 
             var sink = new WallpaperSink(new WallpaperSinkOptions(), desktopMock.Object);
             sink.Consume(filePath);
+
+            desktopMock.Verify();
+        }
+
+        [Test]
+        public void Should_set_background_per_screen()
+        {
+            var input = JObject.Parse(@"{""0"": ""C:\\absolute\\path\\batman.jpg"", ""1"": ""C:\\absolute\\path\\joker.jpg""}");
+
+            var expectedFilePathByScreen = new Dictionary<int, string>()
+            {
+                { 0, @"C:\absolute\path\batman.jpg" },
+                { 1, @"C:\absolute\path\joker.jpg" },
+            };
+
+            var desktopMock = new Mock<IDesktop>();
+            desktopMock.Setup(it => it.SetBackgrounds(expectedFilePathByScreen)).Verifiable();
+
+            var sink = new WallpaperSink(new WallpaperSinkOptions(), desktopMock.Object);
+            sink.Consume(input);
 
             desktopMock.Verify();
         }
@@ -57,6 +78,27 @@ namespace AllMyLights.Test
             var expected = Path.Combine(relativeTo, filePath);
 
             desktopMock.Verify(it => it.SetBackground(expected));
+        }
+
+        [Test]
+        public void Should_prepend_configured_directory_to_relative_paths_for_each_display()
+        {
+            var input = JObject.Parse(@"{""0"": ""relative\\path\\batman.jpg"", ""1"": ""relative\\path\\alfred.jpg""}");
+
+            var expectedFilePathByScreen = new Dictionary<int, string>()
+            {
+                { 0, @"D:\wayne\enterprises\relative\path\batman.jpg" },
+                { 1, @"D:\wayne\enterprises\relative\path\alfred.jpg" },
+            };
+
+            var desktopMock = new Mock<IDesktop>();
+            desktopMock.Setup(it => it.SetBackgrounds(expectedFilePathByScreen)).Verifiable();
+
+            var relativeTo = @"D:\wayne\enterprises";
+            var sink = new WallpaperSink(new WallpaperSinkOptions() { RelativeTo = relativeTo }, desktopMock.Object);
+            sink.Consume(input);
+
+            desktopMock.Verify();
         }
     }
 }