+++ AllMyLights/Platforms/IDesktop.cs
@@ -1,7 +1,11 @@
-﻿namespace AllMyLights.Platforms
+﻿using System.Collections.Generic;
+
+namespace AllMyLights.Platforms
 {
     public interface IDesktop
     {
         void SetBackground(string filePath);
+        void SetBackgrounds(Dictionary<int, string> filePathByScreen);
+        IEnumerable<int> GetScreens();
     }
 }