+++ AllMyLights/Transformations/Mapping/MappingTransformation.cs
@@ -7,22 +7,22 @@ using NLog;
 
 namespace AllMyLights.Transformations.Mapping
 {
-    public class MappingTransformation : ITransformation<string>
+    public class MappingTransformation : ITransformation<object>
     {
         private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
 
         public const string Type = "Mapping";
 
-        private IEnumerable<Tuple<Regex, string>> Mappings { get; }
+        private IEnumerable<Tuple<Regex, object>> Mappings { get; }
         private bool FailOnMiss { get; } = false;
 
         public MappingTransformation(MappingTransformationOptions options)
         {
-            Mappings = options.Mappings.Select(t => new Tuple<Regex, string>(new Regex(t.From), t.To));
+            Mappings = options.Mappings.Select(t => new Tuple<Regex, object>(new Regex(t.From), t.To));
             FailOnMiss = options.FailOnMiss;
         }
 
-        public Func<IObservable<object>, IObservable<string>> GetOperator()
+        public Func<IObservable<object>, IObservable<object>> GetOperator()
         {
             return (source) => source.Select(input =>
             {
@@ -39,12 +39,18 @@ namespace AllMyLights.Transformations.Mapping
 
                 Logger.Debug(() => $"{(match.Success ? $"{inputString} matched by {mapping.Item1}. Substituting {mapping.Item2}..." : "No applicable mapping found.")}");
 
-                IObservable<string> output = match.Success ?
-                    Observable.Return(match.Result(mapping.Item2)) :
-                    (FailOnMiss ? Observable.Empty<string>() : Observable.Return(inputString));
+                if (!match.Success)
+                {
+                    return FailOnMiss ? Observable.Empty<string>() : Observable.Return(inputString);
+                }
+
+                if(mapping.Item2 is string to)
+                {
+                    return Observable.Return(match.Result(to));
+                }
 
 
-                return output;
+                return Observable.Return(mapping.Item2);
             })
             .Switch()
             .Do((output) =>